fix: put all parsing and stringification into the wallet ui
