+++ packages/dapp-svelte-wallet/api/package.json
@@ -27,8 +27,7 @@
   "dependencies": {
     "@agoric/eventual-send": "^0.12.0-dev.0",
     "@agoric/store": "^0.3.1-dev.0",
-    "esm": "^3.2.5",
-    "json5": "^2.1.3"
+    "esm": "^3.2.5"
   },
   "eslintConfig": {
     "extends": [
