+++ packages/dapp-svelte-wallet/ui/package.json
@@ -16,6 +16,7 @@
     "@agoric/captp": "^1.5.2-dev.0",
     "@rollup/plugin-commonjs": "^12.0.0",
     "@rollup/plugin-node-resolve": "^8.0.0",
+    "json5": "^2.1.3",
     "livereload-js": "https://github.com/agoric-labs/livereload-js",
     "rollup": "^2.3.4",
     "rollup-plugin-livereload": "^1.0.0",
