+++ packages/dapp-svelte-wallet/ui/src/Amount.svelte
@@ -1,39 +1,14 @@
 <script>
   import Petname from "./Petname.svelte";
-  import Debug from "../lib/Debug.svelte";
   import Tooltip from "smelte/src/components/Tooltip";
+  import { stringifyValue } from './display';
 
   export let amount;
   export let displayInfo;
 
-  const CONVENTIONAL_DECIMAL_PLACES = 2;
-
   // The amount gets updated. Make this dynamic
   $: ({ brand, value } = amount);
-  const decimate = v => {
-    if (Array.isArray(v)) {
-      return `${v.length}`;
-    }
-    const { decimalPlaces = 0 } = displayInfo || {};
-
-    const bScale = BigInt(10) ** BigInt(decimalPlaces);
-    const bValue = BigInt(value);
-    const bDecimals = BigInt(bValue % bScale)
-    const bUnits = bValue / bScale;
-
-    // Convert 100 to '0000100'.
-    const dec0str0 = `${bDecimals}`.padStart(decimalPlaces, '0');
-    const dec0str = dec0str0.replace(/0+$/, '');
-
-    const decstr = dec0str.padEnd(CONVENTIONAL_DECIMAL_PLACES, '0');
-    const unitstr = `${bUnits}`;
-    if (!decstr) {
-      // No decimals to display.
-      return unitstr;
-    }
-
-    return `${unitstr}.${decstr}`;
-  };
+  const stringify = v => stringifyValue(v, displayInfo);
 </script>
 
 <style>
@@ -63,7 +38,7 @@
     </Tooltip>
   {:else}
     <b>
-      {decimate(value)}
+      {stringify(value)}
       <Petname name={brand.petname} />
     </b>
   {/if}
